refactor: 환불 작업 도메인 분리
