+++ cherrypic-api/src/main/java/org/cherrypic/domain/refundtask/event/RefundScheduleEventListener.java
@@ -1,12 +1,12 @@
-package org.cherrypic.domain.payment.event;
+package org.cherrypic.domain.refundtask.event;
 
 import java.time.Instant;
 import java.time.ZoneId;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import org.cherrypic.domain.payment.dto.RefundTaskDto;
 import org.cherrypic.domain.payment.dto.event.PaymentVerifyEvent;
-import org.cherrypic.domain.payment.service.PaymentAutoRefundService;
+import org.cherrypic.domain.refundtask.dto.RefundTaskDto;
+import org.cherrypic.domain.refundtask.service.RefundTaskService;
 import org.springframework.boot.context.event.ApplicationReadyEvent;
 import org.springframework.context.event.EventListener;
 import org.springframework.scheduling.TaskScheduler;
@@ -19,24 +19,24 @@ import org.springframework.transaction.event.TransactionalEventListener;
 public class RefundScheduleEventListener {
 
     private final TaskScheduler taskScheduler;
-    private final PaymentAutoRefundService paymentAutoRefundService;
+    private final RefundTaskService refundTaskService;
 
     @EventListener(ApplicationReadyEvent.class)
     public void initRefundTaskSchedule() {
-        List<RefundTaskDto> pendingTasks = paymentAutoRefundService.findAllPending();
+        List<RefundTaskDto> pendingTasks = refundTaskService.findAllPending();
         pendingTasks.forEach(this::addSchedule);
     }
 
     @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
     public void addRefundTaskSchedule(PaymentVerifyEvent event) {
         RefundTaskDto refundTask =
-                paymentAutoRefundService.createRefundTask(event.paymentId(), event.paidAt());
+                refundTaskService.createRefundTask(event.paymentId(), event.paidAt());
         addSchedule(refundTask);
     }
 
     private void addSchedule(RefundTaskDto refundTask) {
         Long paymentId = refundTask.paymentId();
         Instant scheduledAt = refundTask.scheduledAt().atZone(ZoneId.systemDefault()).toInstant();
-        taskScheduler.schedule(() -> paymentAutoRefundService.refund(paymentId), scheduledAt);
+        taskScheduler.schedule(() -> refundTaskService.refundPayment(paymentId), scheduledAt);
     }
 }
